+++ backend/src/main/java/com/food/backend/repository/RecipeRepository.java
@@ -1,5 +1,6 @@
 package com.food.backend.repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -20,7 +21,7 @@ public interface RecipeRepository extends JpaRepository<Recipe, Integer> {
     List<Ingredients> findAllIngredientsByRecipeId(@Param("id") Integer id);
 
     @Query(value = "INSERT INTO recipe_to_ingredient(recipe_id, ingredient_id, quantity) VALUES(:id_recipe,:id_ingredient, :quantity)", nativeQuery = true)
-    void addAllIngredientToRecipe(@Param("id_recipe")  Integer recipeId, @Param("id_ingredient") Integer ingredientId, @Param("quantity") Integer quantity);
+    void addAllIngredientToRecipe(@Param("id_recipe")  Integer recipeId, @Param("id_ingredient") Integer ingredientId, @Param("quantity") BigDecimal quantity);
 
     @Modifying
     @Transactional
